+++ packages/contracts/contracts/voting/multisig/Multisig.sol
@@ -20,14 +20,14 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
 
     /// @notice A container for proposal-related information.
     /// @param executed Whether the proposal is executed or not.
+    /// @param approvals The number of approvals casted.
     /// @param parameters The proposal-specific approve settings at the time of the proposal creation.
-    /// @param tally The approve tally of the proposal.
     /// @param approvers The approves casted by the approvers.
     /// @param actions The actions to be executed when the proposal passes.
     struct Proposal {
         bool executed;
+        uint32 approvals;
         ProposalParameters parameters;
-        Tally tally;
         mapping(address => bool) approvers;
         IDAO.Action[] actions;
     }
@@ -38,18 +38,10 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
     /// @param startDate The timestamp when the proposal starts.
     /// @param endDate The timestamp when the proposal expires.
     struct ProposalParameters {
-        uint256 minApprovals;
+        uint16 minApprovals;
         uint64 snapshotBlock;
         uint64 startDate;
         uint64 endDate;
-    }
-
-    /// @notice A container for the proposal tally.
-    /// @param approvals The number of approvals casted.
-    /// @param addresslistLength The length of the addresslist.
-    struct Tally {
-        uint256 approvals;
-        uint256 addresslistLength;
     }
 
     /// @notice A container for the plugin settings.
@@ -99,7 +91,7 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
     /// @notice Thrown if the minimal approvals value is out of bounds (less than 1 or greater than the number of members in the address list).
     /// @param limit The maximal value.
     /// @param actual The actual value.
-    error MinApprovalsOutOfBounds(uint256 limit, uint256 actual);
+    error MinApprovalsOutOfBounds(uint256 limit, uint16 actual);
 
     /// @notice Thrown if the start date is to small.
     /// @param limit The limit value.
@@ -153,8 +145,8 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
     /// @notice Returns the number of approvals,
     /// @param _proposalId The ID of the proposal.
     /// @return The number of approvals.
-    function approvals(uint256 _proposalId) public view returns (uint256) {
-        return proposals[_proposalId].tally.approvals;
+    function approvals(uint256 _proposalId) public view returns (uint32) {
+        return proposals[_proposalId].approvals;
     }
 
     /// @notice Adds new members to the address list and updates the minimum approval parameter.
@@ -174,7 +166,7 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
 
         // Check if the new address list has become shorter than the current minimum number of approvals required.
         uint256 newAddresslistLength = addresslistLength();
-        uint256 minApprovals_ = multisigSettings.minApprovals;
+        uint16 minApprovals_ = multisigSettings.minApprovals;
         if (newAddresslistLength < minApprovals_) {
             revert MinApprovalsOutOfBounds({limit: newAddresslistLength, actual: minApprovals_});
         }
@@ -230,10 +222,9 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
         Proposal storage proposal_ = proposals[proposalId];
 
         proposal_.parameters.snapshotBlock = snapshotBlock;
-        proposal_.parameters.minApprovals = multisigSettings.minApprovals;
         proposal_.parameters.startDate = _startDate;
         proposal_.parameters.endDate = _endDate;
-        proposal_.tally.addresslistLength = addresslistLengthAtBlock(snapshotBlock);
+        proposal_.parameters.minApprovals = multisigSettings.minApprovals;
 
         for (uint256 i; i < _actions.length; ) {
             proposal_.actions.push(_actions[i]);
@@ -257,10 +248,9 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
         }
 
         Proposal storage proposal_ = proposals[_proposalId];
-
-        unchecked {
-            proposal_.tally.approvals += 1;
-        }
+        
+        // should throw an error if more people approve than uint32.
+        proposal_.approvals += 1;
 
         proposal_.approvers[approver] = true;
 
@@ -292,8 +282,8 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
     /// @notice Returns all information for a proposal vote by its ID.
     /// @param _proposalId The ID of the proposal.
     /// @return executed Whether the proposal is executed or not.
+    /// @return approvals The number of approvals casted.
     /// @return parameters The parameters of the proposal vote.
-    /// @return tally The current tally of the proposal vote.
     /// @return actions The actions to be executed in the associated DAO after the proposal has passed.
     function getProposal(
         uint256 _proposalId
@@ -302,16 +292,16 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
         view
         returns (
             bool executed,
+            uint32 approvals,
             ProposalParameters memory parameters,
-            Tally memory tally,
             IDAO.Action[] memory actions
         )
     {
         Proposal storage proposal_ = proposals[_proposalId];
 
         executed = proposal_.executed;
+        approvals = proposal_.approvals;
         parameters = proposal_.parameters;
-        tally = proposal_.tally;
         actions = proposal_.actions;
     }
 
@@ -379,7 +369,7 @@ contract Multisig is PluginUUPSUpgradeable, ProposalUpgradeable, Addresslist {
             return false;
         }
 
-        return proposal_.tally.approvals >= proposal_.parameters.minApprovals;
+        return proposal_.approvals >= proposal_.parameters.minApprovals;
     }
 
     /// @notice Internal function to check if a proposal vote is still open.